LayoutTests/fast/events/spatial-navigation/snav-unit-overflow-and-scroll-in-direction-expected.txt

Start
End
PASS document.activeElement.getAttribute("id") is "start"
PASS true is true
PASS document.activeElement.getAttribute("id") is "end"
